Cleaned up lightkurve based python code.
Editing ipynb's directly in github may need further investigation.  Uploading the py code in meantime.
Side note on local instance of matlib.pyplot (not completely resolved - but not sure if we need to spend more time)

Trying to undestand why and how local pyplot instance is needed/used:
------------
1. lightkurve's plot() method returns an axes object, that is to be displayed through show() method.
2. lightkurve also has a local instance of pyplot
3. lightkurve does not itself call show(), and is upto client app to call show().
4. Unanswered q: how is axes object returned by p)lot() method visible to client instance of pyplot?

Alterantives I tried and work:
1. Modified kightkurve's plot() method to set title and to show it's generated axes data.  This is available (local) at /Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages/lightkurve/lightcurve.py.plot+show
2. Removed calls to pyplot.title() and pyplot.show() from lightkurve client aopp.
However, above seems unnecessary, and we should go with code as is.